fix: typefix

+++ src/lez_core_module.cpp
@@ -1278,8 +1278,8 @@ std::string LEZCoreModule::get_sequencer_addr() {
     return value;
 }
 
-uintptr_t LEZCoreModule::get_distribution_limit() {
-    uintptr_t distribution_limit = 0;
+uint64_t LEZCoreModule::get_distribution_limit() {
+    uint64_t distribution_limit = 0;
     const WalletFfiError error = wallet_ffi_get_distribution_limit(walletHandle, &distribution_limit);
     if (error != SUCCESS) {
         fprintf(stderr, "wallet_ffi_get_distribution_limit: wallet FFI error %d\n", error);
@@ -1289,8 +1289,8 @@ uintptr_t LEZCoreModule::get_distribution_limit() {
     return distribution_limit;
 }
 
-uintptr_t LEZCoreModule::get_callibration_limit() {
-    uintptr_t callibration_limit = 0;
+uint64_t LEZCoreModule::get_callibration_limit() {
+    uint64_t callibration_limit = 0;
     const WalletFfiError error = wallet_ffi_get_callibration_limit(walletHandle, &callibration_limit);
     if (error != SUCCESS) {
         fprintf(stderr, "wallet_ffi_get_callibration_limit: wallet FFI error %d\n", error);
@@ -1300,7 +1300,7 @@ uintptr_t LEZCoreModule::get_callibration_limit() {
     return callibration_limit;
 }
 
-int64_t LEZCoreModule::set_callibration_limit(uintptr_t callibration_limit) {
+int64_t LEZCoreModule::set_callibration_limit(uint64_t callibration_limit) {
     const WalletFfiError error = wallet_ffi_set_distribution_limit(walletHandle, callibration_limit);
     if (error != SUCCESS) {
         fprintf(stderr, "wallet_ffi_set_callibration_limit: wallet FFI error %d\n", error);
@@ -1310,7 +1310,7 @@ int64_t LEZCoreModule::set_callibration_limit(uintptr_t callibration_limit) {
     return SUCCESS;
 }
 
-int64_t LEZCoreModule::set_distribution_limit(uintptr_t distribution_limit) {
+int64_t LEZCoreModule::set_distribution_limit(uint64_t distribution_limit) {
     const WalletFfiError error = wallet_ffi_set_distribution_limit(walletHandle, distribution_limit);
     if (error != SUCCESS) {
         fprintf(stderr, "wallet_ffi_set_distribution_limit: wallet FFI error %d\n", error);
